Make ExceptionNotification work even when ShowExceptions is not present.
In Rails 3.0.x, the ShowExceptions middleware is now responsible for converting
404s into exceptions.  This is important if you use the ExceptionNotifier
middleware with ignore_exceptions => {} (to see 404's, for instance).  Problem
is, it only works if you have show_exceptions enabled (thus the middleware is
included).

Now we make sure that non-exception 404's raise the exception equivalent, at
least locally.  This still allows ignore_exceptions a shot at ignoring the
event, while producing a consistent behaviour for 404's: if I use the middleware
and don't include 404-equivalents in :ignore_exceptions, I want to be notified,
period.

We just have to make sure that if we didn't trigger on an exception to begin
with, not to propagate it past ourselves.